Fri 19 Aug 2016 17:08 commit by king
SQL : Controller's coordinate changed from int to double (0.0 to 1.0)

Functionality : Added Sensor modeling

+++ sql/InitializationHSQL/CreateTables.sql
@@ -6,8 +6,8 @@ CREATE TABLE Site(
 CREATE TABLE Controller(
 	ControllerName varchar(100) PRIMARY KEY,
 	Site varchar(100),
-	PositionX int,
-	PositionY int,
+	PositionX double,
+	PositionY double,
 	FOREIGN KEY (Site) REFERENCES Site(SiteName)
 )
 @
